+++ androidtestapp/src/main/java/com/example/androidtestapp/MainActivity.kt
@@ -294,7 +294,6 @@ class MainActivity : ComponentActivity(), KoinComponent, CoroutineScope by Corou
                     }
                 } else if (isCatalogPage) {
                     var catalog = Catalog(this@MainActivity)
-                    catalog.bind("62", "test")
                     catalog.Content()
                 } else {
                     Column(